Re-evaluate possible overflow points
SafeInt is now used in Diff to prevent undefined behavior should addition create an out of bound value.
Not expected but unable to easily rule this out given the presence of arbitrary input files.
KDiff3App::m_neededLines is shielded but checks on the size of m_diff3LineList.

+++ src/diff.cpp
@@ -647,8 +647,8 @@ void DiffList::runDiff(const std::shared_ptr<LineDataVector> &p1, const size_t i
             Diff d((LineCount)(e->line0 - currentLine1), e->deleted, e->inserted);
             assert(d.numberOfEquals() == e->line1 - currentLine2);
 
-            currentLine1 += (LineRef)(d.numberOfEquals() + d.diff1());
-            currentLine2 += (LineRef)(d.numberOfEquals() + d.diff2());
+            currentLine1 += LineRef((qint64)d.numberOfEquals() + d.diff1());
+            currentLine2 += LineRef((qint64)d.numberOfEquals() + d.diff2());
             push_back(d);
 
             p = e->link;
